requirements.txt

python-dotenv 
requests
cachetools
pillow
aiohttp
google-cloud-texttospeech 
google-cloud-speech  
google.generativeai
Anthropic
langdetect 
keyring 
sounddevice 
soundfile 
pygame 
tkfontchooser 
textblob
pytz 
customtkinter
python-VLC
func_timeout
pyspellchecker
numpy